refactored to put fetch details into the movie details so the page would function on reload
